refactor(codegen, syntax): move `LineTerminatorSplitter` from codegen to syntax crate (#16484)
`LineTerminatorSplitter` is useful in formatter, so move it to syntax, so that we can reuse in the other crates
