+++ crates/oxc_codegen/src/sourcemap_builder.rs
@@ -2,9 +2,10 @@ use std::path::Path;
 
 use oxc_index::{IndexVec, define_nonmax_u32_index_type};
 use oxc_span::Span;
-use oxc_syntax::identifier::{LS, PS};
-
-use crate::str::{LS_LAST_2_BYTES, LS_OR_PS_FIRST_BYTE, PS_LAST_2_BYTES};
+use oxc_syntax::{
+    identifier::{LS, PS},
+    line_terminator::{LS_LAST_2_BYTES, LS_OR_PS_FIRST_BYTE, PS_LAST_2_BYTES},
+};
 
 /// Number of lines to check with linear search when translating byte position to line index
 const LINE_SEARCH_LINEAR_ITERATIONS: usize = 16;
